requirements.txt
pip~=21.3.1
wheel~=0.37.0
google~=3.0.0
tornado~=6.1
setuptools~=56.0.0
future~=0.18.2
streamlit~=1.2.0
nltk~=3.6.5
requests~=2.26.0
numpy~=1.19.5
beautifulsoup4~=4.10.0
pytesseract~=0.3.8
pdf2image~=1.16.0
Pillow~=8.3.2